+++ src/main/java/com/example/auction/domain/category/controller/CategoryController.java
@@ -3,7 +3,6 @@ package com.example.auction.domain.category.controller;
 import com.example.auction.common.dto.BaseResponse;
 import com.example.auction.domain.category.dto.*;
 import com.example.auction.domain.category.service.CategoryService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,32 +15,6 @@ import java.util.List;
 public class CategoryController {
 
     private final CategoryService categoryService;
-
-    @PostMapping("/api/admin/categories")
-    public ResponseEntity<BaseResponse<CategoryCreateResponse>> createCategory(
-            @Valid @RequestBody CategoryCreateRequest request
-    ) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(BaseResponse.success(
-                HttpStatus.CREATED.name(), "카테고리 생성 요청 성공", categoryService.createCategory(request)));
-    }
-
-    @PatchMapping("/api/admin/categories/{categoryId}/name")
-    public ResponseEntity<BaseResponse<CategoryRenameResponse>> renameCategory(
-            @PathVariable Long categoryId,
-            @Valid @RequestBody CategoryRenameRequest request
-    ) {
-        return ResponseEntity.status(HttpStatus.OK).body(BaseResponse.success(
-                HttpStatus.OK.name(), "카테고리 이름 수정 요청 성공", categoryService.renameCategory(categoryId, request)));
-    }
-
-    @PatchMapping("/api/admin/categories/{categoryId}/parent")
-    public ResponseEntity<BaseResponse<CategoryMoveResponse>> moveCategory(
-            @PathVariable Long categoryId,
-            @Valid @RequestBody CategoryMoveRequest request
-    ) {
-        return ResponseEntity.status(HttpStatus.OK).body(BaseResponse.success(
-                HttpStatus.OK.name(), "카테고리 이동 요청 성공", categoryService.moveCategory(categoryId, request)));
-    }
 
     @GetMapping("/api/categories")
     public ResponseEntity<BaseResponse<List<CategoryListGetResponse>>> getCategoryList() {
